Added Stamina bar

+++ Assets/Script/Status/HealthBar2.cs
@@ -10,7 +10,7 @@ public class HealthBar2 : MonoBehaviour
 	public Gradient gradient;
 	public Image fill;
 
-	public void SetMaxHealth(float health)
+	public void SetMaxHealth(int health)
 	{
 		slider.maxValue = health;
 		slider.value = health;
@@ -18,7 +18,7 @@ public class HealthBar2 : MonoBehaviour
 		fill.color = gradient.Evaluate(1f);
 	}
 
-    public void SetHealth(float health)
+    public void SetHealth(int health)
 	{
 		slider.value = health;
